requirements.txt
backcall==0.2.0
decorator==4.4.2
fake-factory==9999.9.9
Faker==5.0.2
faker-e164==0.1.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
phonenumberslite==8.12.15
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.3
pymongo==3.11.2
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
traitlets==4.3.3
wcwidth==0.2.5
